+++ app/src/main/java/com/solux/moro/ui/followlist/component/followRequestUserItem.kt
@@ -1,4 +1,4 @@
-package com.solux.moro.ui.followList.component
+package com.solux.moro.ui.followlist.component
 
 import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.Image
@@ -10,13 +10,9 @@ import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Close
 import androidx.compose.material3.Button
 import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.HorizontalDivider
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
@@ -26,23 +22,26 @@ import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Devices
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.solux.moro.R
+import com.solux.moro.core.designsystem.theme.MoroPalette
+import com.solux.moro.core.designsystem.theme.MoroTheme
+import com.solux.moro.core.designsystem.theme.MoroThemeType
 import com.solux.moro.data.model.User
+import com.solux.moro.data.model.UserColorPalette
 import com.solux.moro.data.model.UserStats
 import com.solux.moro.ui.profile.component.toPxDp
 import com.solux.moro.ui.profile.component.toPxSp
 
 @Composable
-fun FollowUserItem(
+fun FollowRequestUserItem(
     user: User,
     stats: UserStats,
-    isFollowTab: Boolean,
-    onActionClick: () -> Unit,
-    ) {
+    onAcceptClick: () -> Unit,
+    onDeclineClick: () -> Unit
+) {
 
     Column() {
         Row(
@@ -78,79 +77,94 @@ fun FollowUserItem(
                         .height(65.toPxDp)
                 )
             }
-            if(isFollowTab) {
-                FollowButton(onActionClick)
-            }
-            else{
-                FollowingButton(stats.isFollowing, onActionClick)
+                FollowRequestButton(onAcceptClick, onDeclineClick)
             }
         }
-    }
-    HorizontalDivider(thickness = (2.88 ).toPxDp, color = Color(0xFF262626))
+    HorizontalDivider(thickness = 1.dp, color = Color(0xFF262626))
 }
 
 @Composable
-private fun FollowButton(
-    onFollowClick: () -> Unit
-) {
-    IconButton(onClick = onFollowClick) {
-        Icon(
-            imageVector = Icons.Default.Close,
-            contentDescription = "Remove Follower",
-            tint = Color(0xFFF2F2F2)
-        )
-    }
-}
-@Composable
-private fun FollowingButton(
-    following: Boolean,
-    onFollowingClick: () -> Unit
+private fun FollowRequestButton(
+    onAcceptClick:  () -> Unit,
+    onDeclineClick: () -> Unit
 ){
     var containerColor = Color(0xFF262626)
     var contentColor = Color(0xFFF2F2F2)
     var borderColor=Color(0xFF404040)
-    var text="Following"
-
-    if(following){
-        containerColor = Color(0xFF262626)
-        contentColor = Color(0xFFF2F2F2)
-        borderColor=Color(0xFF404040)
-        text="Following"
-    }
-    else{
-        containerColor =  Color(0xFFF2F2F2)
-        contentColor = Color(0xFF121212)
-        borderColor= Color(0xFF262626)
-        text="Follow"
-    }
+    var text="확인"
+Row(
+    horizontalArrangement = Arrangement.spacedBy(
+        5.dp)
+) {
     Button(
-        onClick = onFollowingClick,
+        onClick = onAcceptClick,
         modifier = Modifier
-            .width(259.20001.toPxDp)
-            .height(97.92.toPxDp), colors = ButtonDefaults.buttonColors(
-            containerColor = containerColor,
-            contentColor = contentColor,
+            .width(55.dp)
+            .height(35.dp), colors = ButtonDefaults.buttonColors(
+            containerColor = Color(0xFFF2F2F2),
+            contentColor = Color(0xFF121212),
         ), shape = RoundedCornerShape(size = 23.04.toPxDp),
         contentPadding = PaddingValues(
             horizontal = 0.dp,
             vertical = 0.dp
-        )
-        , border = BorderStroke(width = 2.88.toPxDp, color =borderColor)
+        ), border = BorderStroke(width = 2.88.toPxDp, color = Color(0xFF262626))
     ) {
         Text(
-            text = text,
-            style = TextStyle(
-                fontSize = 40.32.toPxSp,
-                fontWeight = FontWeight(400),
-                textAlign = TextAlign.Center,
-            )
+            text = "확인",
+            style = MoroTheme.typography.bodyRegular16
+        )
+    }
+
+    Button(
+        onClick = onDeclineClick,
+        modifier = Modifier
+            .width(55.dp)
+            .height(35.dp), colors = ButtonDefaults.buttonColors(
+            containerColor = Color(0xFF262626),
+            contentColor = Color(0xFFF2F2F2),
+        ), shape = RoundedCornerShape(size = 23.04.toPxDp),
+        contentPadding = PaddingValues(
+            horizontal = 0.dp,
+            vertical = 0.dp
+        ), border = BorderStroke(width = 2.88.toPxDp, color = Color(0xFF404040))
+    ) {
+        Text(
+            text = "삭제",
+            style = MoroTheme.typography.bodyRegular16
         )
     }
 }
-
+}
 
 @Preview( device = Devices.PIXEL_4A)
 @Composable
-fun FollowUserItemPreview(){
-    //FollowUserItem()
+fun FollowRequestUserItemPreview(){
+    FollowRequestUserItem(
+        user = User(
+            id = 1,
+            email = "test@test.com",
+            nickname = "테스트유저",
+            colorPalette = UserColorPalette(
+                theme = MoroThemeType.Pastel,
+                userColor = MoroPalette.Pastel.Purple400,
+                paletteColors = listOf(
+                    MoroPalette.Pastel.Purple400,
+                    MoroPalette.Pastel.Yellow300,
+                    MoroPalette.Pastel.Green200,
+                    MoroPalette.Pastel.Cyan200,
+                    MoroPalette.Pastel.Indigo500,
+                    MoroPalette.Pastel.Gray400
+                )
+            )
+        ),
+        stats = UserStats(
+            colorsCount = 1,
+            followerCount = 1,
+            followingCount = 1,
+            isFollowing = true
+        ),
+        onAcceptClick = {},
+        onDeclineClick = {}
+    )
 }
+
