+++ app/src/main/java/com/solux/moro/ui/paletteedit/PaletteEditViewModel.kt
@@ -1,4 +1,4 @@
-package com.solux.moro.ui.paletteEdit
+package com.solux.moro.ui.paletteedit
 
 import androidx.compose.ui.graphics.Color
 import androidx.lifecycle.ViewModel
